pyproject.toml
[tool.poetry]
name = "test-task-grigoryev"
version = "0.1.0"
description = ""
authors = ["Sergey <gladkov.sergey87@xmail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
